Added implode function to tables
When calling a table multiple times, you can now modify the output using the implode modifier.
To keep the program simple and versatile, implode requires a quote delimited modifier that shows exactly what character or group of characters should be used with the implode command. The program looks for whatever is inside of the quotes when using the implode command.
For example, [@3 Table >> implode ", "] = run the table 3 times and display the results with a comma followed by an blank space.