Instead of reinventing the wheel, learn why it’s round, then decide whether you really need a square one.
Getting HTML support was getting to be so difficult, I instead decided to take a step back and think about it more. While doing so, I realized I don't need to recreate all the Python formatting rules for HTML for my simple program in order to get HTML working... I just need to send my output to a program ALREADY made for displaying HTML!

Presenting the "Generate to Browser" output button. This sends the output of your table directly to your web browser, allowing it to parse and display all that difficult HTML.

This program can now say it fully supports HTML, because it is standing on the back of a program that fully supports HTML.