Tweaking the recursion and loop break sensitivities
If and ifnot are governed by recursion. This seems to be working great, so I lowered the sensitivities to allow for more recursion.

While and whilenot are governed by loops in place, and because those loops seem to be lag/crashing the program I tweaked it so that it should either run, or error out almost immediately, hopefully without crashing the program.